Adding dockerfile and yaml for deployments
running "make container" will build and tag executor and controller
images. controller is deployed as Deployment config, and executor is a
job that the controller creates in cluster.

still missing - how to pass the secret or config file to communicate
with cluster. we need cert to allow creating nodefence crd

Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>