Using secrets to pass fence method password
When executing the agent pod we add env var for secret name - when agent
is called with --password-script=/usr/sbin/fetch-passwd parameter, the
agent will execute the script to get the password. That way the password
is saved encrypted in secret object with access only to agent pods.

Later patch will create specific service account for each secret, such
that only the agent pod that should read this password will have access
to it.

Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>